Add files via upload
This library template file contains the source and dest locations of the first set of primary amines, secondary amines, carboxylic acids, and aldehydes that were acquired from Enamine in the Summer of 2020.  This version does not contain source information for control compounds to go in the designated control wells, which are located in columns 46, 47, and 48 of the 1536-well assembly plates.